requirements.txt
click==8.0.0
Flask==2.0.0
Flask-Cors==3.0.10
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
six==1.16.0
Werkzeug==2.0.0
dnspython
pymongo==4.3.3
pyjwt[crypto]
ipython
python-dotenv
bcrypt